Checkpoint: Run judge on threads
